feat(server): split embedding client (#12809)
