+++ packages/backend/server/src/plugins/copilot/context/index.ts
@@ -1,3 +1,2 @@
-export { CopilotContextDocJob } from './job';
 export { CopilotContextResolver, CopilotContextRootResolver } from './resolver';
 export { CopilotContextService } from './service';
