--- packages/backend/server/src/plugins/copilot/context/utils.ts
@@ -1,11 +1,0 @@
-export class GqlSignal implements AsyncDisposable {
-  readonly abortController = new AbortController();
-
-  get signal() {
-    return this.abortController.signal;
-  }
-
-  async [Symbol.asyncDispose]() {
-    this.abortController.abort();
-  }
-}
