+++ packages/backend/server/src/plugins/copilot/embedding/index.ts
@@ -1,0 +1,4 @@
+export { getEmbeddingClient, MockEmbeddingClient } from './client';
+export { CopilotEmbeddingJob } from './job';
+export type { Chunk, DocFragment } from './types';
+export { EMBEDDING_DIMENSIONS, EmbeddingClient } from './types';
