+++ packages/backend/server/src/plugins/copilot/embedding/types.ts
